khansaama-data-uploader/requirements.txt
pymongo==4.6.1
faker==24.1.0
python-dotenv==1.0.1
bcrypt==4.1.2